requirements.txt
alembic==1.13.1
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
cachelib==0.13.0
cachetools==5.3.3
cbor2==5.6.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.0
cryptography
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.1
django-allauth==65.0.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-sslserver==0.22
django-webauthn==0.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
Faker==26.0.0
filelock==3.15.1
flash==1.0.3
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-PageDown==0.4.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
fsspec==2025.2.0
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
greenlet==3.1.1
gunicorn==22.0.0
h11==0.14.0
httpie==3.2.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
jwcrypto
jwt
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.0.5
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT
pyOpenSSL
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil
python-dotenv
requests
requests-oauthlib
requests-toolbelt
rich
rsa
selenium
setuptools
six
sniffio
sortedcontainers
SQLAlchemy
sqlparse
sympy
torch
torchaudio
torchvision
trio
trio-websocket
triton
typing_extensions
tzdata
uritemplate
urllib3
virtualenv
visitor
waitress
webauthn
webencodings
websocket-client
Werkzeug
wsproto
WTForms
